dags/requirements.txt
apache-airflow==2.6.3
pymongo==4.6.1
psycopg2-binary==2.9.9
